refactor: break SolInputs to its own crate
