+++ crates/sol-macro-input/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "alloy-sol-macro-input"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+exclude.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+dunce = "1.0.4"
+proc-macro2.workspace = true
+syn.workspace = true
+syn-solidity.workspace = true
+
+# json
+alloy-json-abi = { workspace = true, optional = true }
+serde_json = { workspace = true, optional = true }
+quote = { workspace = true, optional = true }
+
+[features]
+json = ["dep:alloy-json-abi", "dep:serde_json", "dep:quote"]
