+++ crates/sol-macro-input/src/json.rs
@@ -1,53 +1,75 @@
+use crate::{SolInput, SolInputKind};
+
 use alloy_json_abi::{ContractObject, JsonAbi, ToSolConfig};
 use proc_macro2::{Ident, TokenStream};
 use quote::{quote, TokenStreamExt};
-use syn::{Attribute, Result};
+use syn::Result;
 
-pub fn expand(name: Ident, json: ContractObject, attrs: Vec<Attribute>) -> Result<TokenStream> {
-    let ContractObject { abi, bytecode, deployed_bytecode } = json;
+impl SolInput {
+    /// Normalize JSON ABI inputs into Sol inputs.
+    pub fn normalize_json(self) -> Result<Self> {
+        if !matches!(self.kind, SolInputKind::Json(..)) {
+            return Ok(self);
+        }
 
-    let mut abi = abi.ok_or_else(|| syn::Error::new(name.span(), "ABI not found in JSON"))?;
-    let sol = abi_to_sol(&name, &mut abi);
-    let sol_interface_tokens = tokens_for_sol(&name, &sol)?;
-    let bytecode = bytecode.map(|bytes| {
-        let s = bytes.to_string();
-        quote!(bytecode = #s,)
-    });
-    let deployed_bytecode = deployed_bytecode.map(|bytes| {
-        let s = bytes.to_string();
-        quote!(deployed_bytecode = #s)
-    });
+        // irrefutable, as we just checked
+        let SolInput {
+            attrs,
+            path,
+            kind: SolInputKind::Json(name, ContractObject { abi, bytecode, deployed_bytecode }),
+        } = self
+        else {
+            panic!()
+        };
 
-    let doc_str = format!(
-        "\n\n\
-Generated by the following Solidity interface...
-```solidity
-{sol}
-```
+        let mut abi = abi.ok_or_else(|| syn::Error::new(name.span(), "ABI not found in JSON"))?;
+        let sol = abi_to_sol(&name, &mut abi);
+        let sol_interface_tokens = tokens_for_sol(&name, &sol)?;
+        let bytecode = bytecode.map(|bytes| {
+            let s = bytes.to_string();
+            quote!(bytecode = #s,)
+        });
+        let deployed_bytecode = deployed_bytecode.map(|bytes| {
+            let s = bytes.to_string();
+            quote!(deployed_bytecode = #s)
+        });
 
-...which was generated by the following JSON ABI:
-```json
-{json_s}
-```",
-        json_s = serde_json::to_string_pretty(&abi).unwrap()
-    );
-    let tokens = quote! {
-        #(#attrs)*
-        #[doc = #doc_str]
-        #[sol(#bytecode #deployed_bytecode)]
-        #sol_interface_tokens
-    };
+        let attrs_iter = attrs.iter();
+        let doc_str = format!(
+            "\n\n\
+                Generated by the following Solidity interface...
+                ```solidity
+                {sol}
+                ```
 
-    let ast = syn::parse2(tokens).map_err(|e| {
-        let msg = format!(
-            "failed to parse ABI-generated tokens into a Solidity AST: {e}.\n\
-             This is a bug. We would appreciate a bug report: \
-             https://github.com/alloy-rs/core/issues/new/choose"
+                ...which was generated by the following JSON ABI:
+                ```json
+                {json_s}
+                ```",
+            json_s = serde_json::to_string_pretty(&abi).unwrap()
         );
-        syn::Error::new(name.span(), msg)
-    })?;
-    crate::expand::expand(ast)
+        let tokens = quote! {
+            #(#attrs_iter)*
+            #[doc = #doc_str]
+            #[sol(#bytecode #deployed_bytecode)]
+            #sol_interface_tokens
+        };
+
+        let ast: ast::File = syn::parse2(tokens).map_err(|e| {
+            let msg = format!(
+                "failed to parse ABI-generated tokens into a Solidity AST: {e}.\n\
+                 This is a bug. We would appreciate a bug report: \
+                 https://github.com/alloy-rs/core/issues/new/choose"
+            );
+            syn::Error::new(name.span(), msg)
+        })?;
+
+        let kind = SolInputKind::Sol(ast);
+        Ok(SolInput { attrs, path, kind })
+    }
 }
+
+// doesn't parse Json
 
 fn abi_to_sol(name: &Ident, abi: &mut JsonAbi) -> String {
     abi.dedup();
