+++ crates/sol-macro/Cargo.toml
@@ -21,13 +21,14 @@ all-features = true
 rustdoc-args = ["--cfg", "docsrs"]
 
 [dependencies]
+alloy-sol-macro-input.workspace = true
+
 syn-solidity = { workspace = true, features = ["visit", "visit-mut"] }
 
 proc-macro2.workspace = true
 quote.workspace = true
 syn = { workspace = true, features = ["extra-traits"] }
 
-dunce = "1.0"
 heck = "0.4"
 hex.workspace = true
 indexmap = "2"
@@ -36,7 +37,6 @@ tiny-keccak = { workspace = true, features = ["keccak"] }
 
 # json
 alloy-json-abi = { workspace = true, optional = true }
-serde_json = { workspace = true, optional = true }
 
 [features]
-json = ["dep:alloy-json-abi", "dep:serde_json"]
+json = ["dep:alloy-json-abi", "alloy-sol-macro-input/json"]
